+++ packages/api-reference/src/components/Content/Tag/Tag.vue
@@ -14,7 +14,7 @@ import {
   SectionHeader,
   SectionHeaderTag,
 } from '@/components/Section'
-import ExternalDocs from '@/features/ExternalDocs/ExternalDocs.vue'
+import { ExternalDocs } from '@/features/external-docs'
 import { SpecificationExtension } from '@/features/specification-extension'
 import { useConfig } from '@/hooks/useConfig'
 import { useNavState } from '@/hooks/useNavState'
