+++ packages/api-reference/src/components/LinkList/index.ts
@@ -1,0 +1,1 @@
+export { default as LinkList } from './LinkList.vue'
