+++ packages/api-reference/src/features/external-docs/ExternalDocs.vue
