+++ packages/api-reference/src/features/external-docs/index.ts
@@ -1,0 +1,1 @@
+export { default as ExternalDocs } from './ExternalDocs.vue'
