+++ packages/api-reference/src/features/info-object/License.vue
