+++ packages/api-reference/src/features/info-object/TermsOfService.vue
