+++ packages/api-reference/src/features/info-object/index.ts
@@ -1,0 +1,3 @@
+export { default as Contact } from './Contact.vue'
+export { default as License } from './License.vue'
+export { default as TermsOfService } from './TermsOfService.vue'
